Separate send token amount into a component
